Moving files back to top level.
